Add files via upload
This project showcases the development of a Retrieval-Augmented Generation (RAG) chatbot that helps data scientists learn Docker concepts. Utilizing LlamaIndex for indexing, Groq for advanced language model interactions, and Llama3 for generating responses, the chatbot offers detailed explanations on Docker, using a comprehensive book as its knowledge base. The system is designed to be easily customizable, allowing users to substitute the Docker book with other resources for different educational topics or databases. This setup not only facilitates learning Docker but also demonstrates the integration of cutting-edge AI technologies for educational purposes.